requirements.txt
fastapi
starlette
starlette-admin
sqlalchemy
sqlalchemy-file
uvicorn

pillow
fasteners
itsdangerous

passlib
bcrypt
pymysql
cryptography
python-dotenv
httpx